+++ src/App.tsx
@@ -1,23 +1,18 @@
-import './App.scss';
-import { Navigation } from './components/Navigation/Navigation';
-import { Navigate, Route, Routes } from 'react-router-dom';
-import { HomePage } from './components/HomePage/HomePage';
-import { PeoplePage } from './components/PeoplePage/PeoplePage';
-
-export const PageNotFound = () => {
-  return <h1 className="title">Page not found</h1>;
-};
+import { Routes, Route, Navigate } from 'react-router-dom';
+import { HomePage } from './components/HomePage';
+import { PeoplePage } from './components/PeoplePage';
+import { PageNotFound } from './components/PageNotFound';
+import { Navbar } from './components/Navbar';
 
 export const App = () => (
   <div data-cy="app">
-    <Navigation />
-
+    <Navbar />
     <main className="section">
       <div className="container">
         <Routes>
           <Route path="/" element={<HomePage />} />
           <Route path="/home" element={<Navigate to="/" replace />} />
-          <Route path="people/" element={<PeoplePage />} />
+          <Route path="/people" element={<PeoplePage />} />
           <Route path="/people/:slug" element={<PeoplePage />} />
           <Route path="*" element={<PageNotFound />} />
         </Routes>
