+++ src/components/HomePage.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const HomePage: React.FC = () => <h1 className="title">Home Page</h1>;
