--- src/components/HomePage/HomePage.tsx
@@ -1,3 +1,0 @@
-export const HomePage = () => {
-  return <h1 className="title">Home Page</h1>;
-};
