+++ src/components/PageNotFound.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const PageNotFound: React.FC = () => (
+  <h1 className="title">Page not found</h1>
+);
