+++ src/components/PeoplePage.tsx
@@ -1,0 +1,39 @@
+import { getPeople } from '../api';
+import { Person } from '../types/Person';
+import { PeopleTable } from '../components/PeopleTable';
+import { Loader } from '../components/Loader';
+import { useParams } from 'react-router-dom';
+import { useState, useEffect } from 'react';
+
+export const PeoplePage = () => {
+  const [people, setPeople] = useState<Person[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(false);
+  const { slug } = useParams<{ slug: string }>();
+
+  useEffect(() => {
+    setLoading(true);
+    getPeople()
+      .then(data => setPeople(data))
+      .catch(() => setError(true))
+      .finally(() => setLoading(false));
+  }, []);
+
+  return (
+    <>
+      <h1 className="title">People Page</h1>
+      {loading && <Loader />}
+      {error && (
+        <p data-cy="peopleLoadingError" className="has-text-danger">
+          Something went wrong
+        </p>
+      )}
+      {!loading && !error && people.length === 0 && (
+        <p data-cy="noPeopleMessage">There are no people on the server</p>
+      )}
+      {!loading && !error && people.length > 0 && (
+        <PeopleTable people={people} selectedSlug={slug} />
+      )}
+    </>
+  );
+};
