+++ src/components/PeopleTable.tsx
@@ -1,0 +1,69 @@
+import React from 'react';
+import { Person } from '../types/Person';
+import { PersonLink } from './PersonLink';
+
+type Props = {
+  people: Person[];
+  selectedSlug?: string;
+};
+
+export const PeopleTable: React.FC<Props> = ({ people, selectedSlug }) => {
+  return (
+    <div className="box table-container">
+      <table
+        data-cy="peopleTable"
+        className="table is-striped is-hoverable is-narrow is-fullwidth"
+      >
+        <thead>
+          <tr>
+            <th>Name</th>
+            <th>Sex</th>
+            <th>Born</th>
+            <th>Died</th>
+            <th>Mother</th>
+            <th>Father</th>
+          </tr>
+        </thead>
+
+        <tbody>
+          {people.map(person => (
+            <tr
+              key={person.slug}
+              data-cy="person"
+              className={
+                person.slug === selectedSlug ? 'has-background-warning' : ''
+              }
+            >
+              <td>
+                <PersonLink person={person} name={person.name} />
+              </td>
+              <td>{person.sex}</td>
+              <td>{person.born}</td>
+              <td>{person.died}</td>
+              <td>
+                {person.motherName ? (
+                  <PersonLink
+                    person={people.find(p => p.name === person.motherName)}
+                    name={person.motherName}
+                  />
+                ) : (
+                  '-'
+                )}
+              </td>
+              <td>
+                {person.fatherName ? (
+                  <PersonLink
+                    person={people.find(p => p.name === person.fatherName)}
+                    name={person.fatherName}
+                  />
+                ) : (
+                  '-'
+                )}
+              </td>
+            </tr>
+          ))}
+        </tbody>
+      </table>
+    </div>
+  );
+};
