--- src/components/PeopleTable/PeopleTable.tsx
@@ -1,62 +1,0 @@
-import classNames from 'classnames';
-import { Person } from '../../types';
-import { PersonLink } from '../PersonLink/PersonLink';
-
-interface Props {
-  peoples: Person[];
-  selectedSlug?: string;
-}
-
-export const PeopleTable: React.FC<Props> = ({ peoples, selectedSlug }) => {
-  return (
-    <table
-      data-cy="peopleTable"
-      className="table is-striped is-hoverable is-narrow is-fullwidth"
-    >
-      <thead>
-        <tr>
-          <th>Name</th>
-          <th>Sex</th>
-          <th>Born</th>
-          <th>Died</th>
-          <th>Mother</th>
-          <th>Father</th>
-        </tr>
-      </thead>
-
-      <tbody>
-        {peoples.map(person => (
-          <tr
-            data-cy="person"
-            key={person.slug}
-            className={classNames({
-              'has-background-warning': selectedSlug === person.slug,
-            })}
-          >
-            <td>
-              <PersonLink personName={person.name} peoples={peoples} />
-            </td>
-
-            <td>{person.sex}</td>
-            <td>{person.born}</td>
-            <td>{person.died}</td>
-            <td>
-              {person.motherName ? (
-                <PersonLink personName={person.motherName} peoples={peoples} />
-              ) : (
-                '-'
-              )}
-            </td>
-            <td>
-              {person.fatherName ? (
-                <PersonLink personName={person.fatherName} peoples={peoples} />
-              ) : (
-                '-'
-              )}
-            </td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  );
-};
